+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.11.1
 - Regrouping files in folders.
+- Added remoteType to package.json.
 
 # 0.11.0
 - The z80-debug adapter has been renamed to "DeZog".